Refactor nav
