+++ src/App.js
@@ -7,20 +7,20 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { Box } from '@mui/material';
 
 // pages
-import Home from './pages/Home';
-import Today from './pages/Today';
-import Tomorrow from './pages/Tomorrow';
-import Completed from './pages/Completed';
-import PastDue from './pages/PastDue';
-import Upcoming from './pages/Upcoming';
+import Home from 'pages/Home';
+import Today from 'pages/Today';
+import Tomorrow from 'pages/Tomorrow';
+import Completed from 'pages/Completed';
+import PastDue from 'pages/PastDue';
+import Upcoming from 'pages/Upcoming';
 
 // components
-import Header from './components/Header';
-import Nav from './components/Nav';
+import Header from 'components/Header';
+import Nav from 'components/nav/Nav';
 
 // util components
-import SaveToStorage from './components/SaveToStorage';
-import PageChange from './components/PageChange';
+import SaveToStorage from 'components/SaveToStorage';
+import PageChange from 'components/PageChange';
 
 const App = () => (
   <Router>
